refactor(app): extract CustomThemeProvider into its own module (C-4)
Move the ~270-line theme provider block out of the 2 100-line god-App.jsx
into imports/ui/CustomThemeProvider.jsx, the first bounded split of the
file. Exports preserved verbatim: CustomThemeProvider, useTheme,
Meteor.useTheme side-effect, Meteor.useLocation/useNavigate/useParams.
getThemeSetting exported (deviation from module-private intent) so the
remaining App function usage at the <meta name="theme-color"> tag is
DRY. App.jsx re-exports all three for backward-compat import paths.
ThemingPage.jsx import { history, useTheme } from './App' unchanged.
Removed now-unused imports: ThemeProvider, CssBaseline, createTheme,
Tracker, createContext, useContext, useRef. TDD smoke test added to
imports/ui/customThemeProvider.tests.js, wired through tests/main.js.
